Fix bug in defuzzer for multi-dimensional tuples
When only jittering twice for the whole tuple, there are adjacent cells
in the grid that can be considered unequal, leading to points
arbitrarily close together that do not map to each other.

Instead, jitter an n-tuple 2**n ways, twice for each dimension
independently.  This properly covers the space around each cell.